% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celcius_to_fahrenheit.R
\name{celcius_to_fahrenheit}
\alias{celcius_to_fahrenheit}
\title{\emph{Celcius To Farenheit}}
\usage{
celcius_to_fahrenheit(temp_centigrados)
}
\arguments{
\item{temp_centigrados}{A numeric value representing the temperature in Celsius.}
}
\value{
A numeric value representing the temperature in Fahrenheit.
}
\description{
This function \code{convertir_C_a_F()}converts a temperature given in Celsius to Fahrenheit.
}
\details{
This are extra details
}
\examples{
convertir_C_a_F(20)

}
