% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{cleanAndSubsetDetailedData}
\alias{cleanAndSubsetDetailedData}
\title{need to find the Format of the Time stamp first.
It could be one of two formats...
If neither, then don't convert..}
\usage{
cleanAndSubsetDetailedData(wxdata, date, opt_temperature_columns = TRUE,
  opt_all_columns = FALSE, opt_custom_columns = FALSE,
  custom_columns = NULL, opt_compress_output = FALSE, opt_verbose = FALSE)
}
\description{
need to find the Format of the Time stamp first.
It could be one of two formats...
If neither, then don't convert..
}

