% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celcius.R
\name{fahrenheit_to_celcius}
\alias{fahrenheit_to_celcius}
\title{\emph{Farenheit To Celcius}}
\usage{
fahrenheit_to_celcius(temperature_fahrenheit)
}
\arguments{
\item{temperature_fahrenheit}{A numeric value representing the temperature in Farenheit.}
}
\value{
A numeric value representing the temperature in Celcius.
}
\description{
This function \code{fahrenheit_to_celcius} converts a temperature given in Farenheit to Celcius.
}
\details{
This are extra details
}
\examples{
fahrenheit_to_celcius(50)

}
