\name{FetchDailyWeatherForStation}
\alias{FetchDailyWeatherForStation}
\title{This function gets data from Wx Underground, given a valid station and date}
\usage{
FetchDailyWeatherForStation(station, date, station_type = "airportCode",
  opt.temp.only = T, opt.compress.output = FALSE, opt.debug = FALSE)
}
\description{
TODO: Error checking for valid station name, date,
station.type
}

