% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{generate_40_year_average_temperature_plot}
\alias{generate_40_year_average_temperature_plot}
\title{\emph{Generate 40-Year Average Monthly Temperature Plot}}
\usage{
generate_40_year_average_temperature_plot(id)
}
\arguments{
\item{id}{Character string representing the ID of the weather station for which to
generate the plot. This should match the station IDs in the \code{weatherdata} dataset.}
}
\value{
A ggplot object displaying the average monthly temperature for the specified
weather station across 40-year intervals.
}
\description{
The function \code{generate_40_year_average_temperature_plot()} generates a plot showing the
average temperature of the specified weather station for each month over 40-year intervals.
This allows for easy comparison of average temperatures across different months for each 40-year
period available in the dataset.
}
\examples{
generate_40_year_average_temperature_plot("NH0472")

}
