% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meteorological_summary.R
\name{get_meteorological_summary}
\alias{get_meteorological_summary}
\title{\emph{Get Meteorological Summary by ID}}
\usage{
get_meteorological_summary(id)
}
\arguments{
\item{id}{A character or numeric value representing the unique identifier
of the weather station.}
}
\value{
A named list with descriptions and values:
\describe{
\item{temp_mean}{Mean temperature in Celsius.}
\item{temp_max}{Maximum temperature in Celsius.}
\item{temp_min}{Minimum temperature in Celsius.}
\item{humidity_mean}{Average relative humidity in percentage.}
}
}
\description{
This function returns key meteorological summaries, such as mean temperature,
maximum temperature, minimum temperature, and average humidity, for a specific \code{id}.
}
\examples{

get_meteorological_summary("NH0046")

}
