% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{IntlWxStations}
\alias{IntlWxStations}
\title{Data - International Weather Stations}
\format{An object of class \code{data.frame} with 9715 rows and 1 columns.}
\usage{
data(IntlWxStations)
}
\description{
This is a data frame of the 1602 stations in Weather Underground's
 database. The 4-letter "ICAO" is used by the functions in this package
 to check and get the weather data. Note that not all the stations
 have weather data.
}
\references{
This data frame has been created by 
\url{http://weather.rap.ucar.edu/surface/stations.txt}
maintained by Greg Thompson of NCAR.
}
\author{
Ram Narasimhan \email{ramnarasimhan@gmail.com}
}
\keyword{data}
