\name{IsStationTypeInvalid}
\alias{IsStationTypeInvalid}
\title{Check if the station type is airportCode or id}
\usage{
IsStationTypeInvalid(station_type)
}
\arguments{
  \item{station_type}{can be \code{airportCode} which is
  the default, or it can be \code{id} which is a
  weather-station ID}
}
\description{
We are checking if a valid station type was given to the
function.
}
\seealso{
getStationCode
}

