% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{London2013}
\alias{London2013}
\title{Data - Ambient Temperature for the City of London for all of 2013}
\usage{
data(London2013)
}
\description{
This is a data frame of Ambient temperature data, extracted from Weather Undergound.
Each row has two entries (columns). The Timestamp (YYYY-MM-DD HH:MM:SS) and the Temperature (in degrees F)
}
\author{
Ram Narasimhan \email{ramnarasimhan@gmail.com}
}
\references{
\url{http://www.wunderground.com/history/airport/EGLL/2013/1/1/DailyHistory.html?format=1}
}
\keyword{data}

