% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{SFO2012}
\alias{SFO2012}
\title{Data - Ambient Temperature for the City of San Francisco for all of 2012}
\usage{
data(SFO2012)
}
\description{
This is a data frame of Ambient temperature data, extracted from Weather Undergound.
Each row has two entries (columns). The Timestamp (YYYY-MM-DD HH:MM:SS) and the Temperature (in degrees F)
}
\author{
Ram Narasimhan \email{ramnarasimhan@gmail.com}
}
\references{
\url{http://www.wunderground.com/history/airport/KSFO/2012/1/1/DailyHistory.html?format=1}
}
\keyword{data}

