% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{SFO2013Summarized}
\alias{SFO2013Summarized}
\title{Data - Summarized Daily Temperature for the City of San Francisco for all of 2013}
\usage{
data(SFO2013Summarized)
}
\description{
This is a data frame of Ambient temperature data, extracted from Weather Undergound.
Each row has four columns. The Timestamp (YYYY-MM-DD HH:MM:SS) and three Temperature Columns: Daily Max, Mean and Min (in degrees F)
In comparison with the \code{SFO2013} dataset which has 9507 rows, this dataset has exactly
365 rows, one for each day in 2013.
}
\author{
Ram Narasimhan \email{ramnarasimhan@gmail.com}
}
\references{
\url{http://www.wunderground.com/history/airport/SFO/2013/1/1/CustomHistory.html?dayend=31&monthend=12&yearend=2013&req_city=NA&req_state=NA&req_statename=NA&format=1}
}
\keyword{data}

