% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{USAirportWeatherStations}
\alias{USAirportWeatherStations}
\title{Data - US Weather Stations ID's}
\format{An object of class \code{data.frame} with 1602 rows and 7 columns.}
\usage{
data(USAirportWeatherStations)
}
\description{
This is a data frame of the 1602 stations in Weather Underground's
 database. The 4-letter "airportCode" is used by functions
 to check and get the weather data.
}
\references{
\url{http://www.wunderground.com/about/faq/US_cities.asp}
}
\author{
Ram Narasimhan \email{ramnarasimhan@gmail.com}
}
\keyword{data}
