% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherdata-package.R
\docType{package}
\name{weatherdata-package}
\alias{weatherdata}
\alias{weatherdata-package}
\title{weatherdata: Structure of a Package for Reading and Analyzing Meteorological Station Data}
\description{
The package is designed to provide tools for efficiently handling meteorological station data. The structure is organized to facilitate modularity, ease of use, and maintainability.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BMilesRouillon/weatherdata}
  \item \url{https://bmilesrouillon.github.io/weatherdata/}
  \item Report bugs at \url{https://github.com/BMilesRouillon/weatherdata/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Miles Rouillon \email{bmiles@mail.austral.edu.ar} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Facundo Rubiolo \email{frubiolo@mail.austral.edu.ar} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
