% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_and_read_weatherdata.R
\name{download_and_read_weatherdata}
\alias{download_and_read_weatherdata}
\title{\emph{Download and Read Weather Data}}
\usage{
download_and_read_weatherdata()
}
\value{
A data frame containing the combined data from multiple meteorological stations.
}
\description{
The function \code{download_and_read_weatherdata()} downloads meteorological data from specified URLs
and combines the datasets into a single data frame. The function reads data directly from the
provided URLs.
}
\details{
This function fetches data from various meteorological stations and combines them for further analysis.
}
\examples{
download_and_read_weatherdata()

}
