\name{weatherData-package}
\alias{weatherData-package}
\alias{weatherData}
\docType{package}
\title{
weatherData: global daily weather data
}
\description{
This package allows the user to derive global daily weather data from a number of sources.
}
\details{
\tabular{ll}{
Package: \tab weatherData\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-07-19\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

}
\author{
Jacob van Etten

Maintainer: Jacob van Etten <jacobvanetten@yahoo.com>
}

\keyword{ package }

